add JSX Typings
