+++ src/createDSComponent.ts
@@ -19,7 +19,7 @@ type DeStagnateConstructor<Props, State> = new (
 /* eslint-disable @typescript-eslint/naming-convention */
 /**
  * Creates nested DeStagnate component
- * @deprecated do not use this function, since 1.7.0
+ * @deprecated do not use this function, since 1.6.1
  * @see {@link https://github.com/Luke-zhang-04/DeStagnate/wiki/createDSComponent}
  * @see {@link https://github.com/Luke-zhang-04/DeStagnate/wiki/nestedComponents}
  * @param {DeStagnateConstructor} Component - DeStagnate component
